pyproject.toml
[project]
name = "cs1430-final-project"
version = "0.1.0"
description = "Automatic sports game annotation — player/ball detection, field reconstruction, and 2D play visualization"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Arib Syed" },
    { name = "Heewon Seo" },
    { name = "Oras Panagam" },
    { name = "Jeffrey Mu" },
]

dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "ultralytics>=8.0.0",
    "supervision>=0.20.0",
    "opencv-python>=4.9.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "pillow>=10.2.0",
    "matplotlib>=3.8.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "pytest>=8.0.0",
    "ruff>=0.3.0",
]

[project.scripts]
cs1430-export-yolo = "cs1430_runner.shim:export_yolo_main"
cs1430-train-yolo = "cs1430_runner.shim:train_yolo_main"
cs1430-annotate = "cs1430_runner.shim:annotate_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["code", "cs1430_runner"]

[dependency-groups]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "pytest>=8.0.0",
    "ruff>=0.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

# torch cu124 wheels (older cluster drivers)
[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }
